Load the `marked.js` library
We want to be able to just load one external file from our HTML
skeleton and have it do all the rest. Let’s have it load the `marked.js`
library.

There is a slight problem: If we use relative paths in Javascript, they
are relative to the HTML page. However, we’d like to eventually host the
whole engine on the Web so presentations could just include it and have
it pull all its dependencies.

To that end, we will require that the `<script>` element that loads us
is marked with a distinctive `id` attribute. Then we find it, take its
path and refer to the library relative to that.