Navigate slides by keys
There are two events that interest us; roughly, `keypress` for those
that depend on the keyboard layout and `keydown` for those that don’t.

We store the key bindings as maps from the key code to the handler
function. If there is a handler for a particular key combination, we
call it.