Set font size
None of the usual font size units are good for presentations. A
comprehensive [research][1] on the matter involves taking into account
the screen size (when actually projected), the room depth (eye distance
of the farthest viewer), angular seconds and the visual acuity of
viewers.

[1]: http://www.thinkoutsidetheslide.com/selecting-the-correct-font-size/

We just base everything on a base size of 24pt on a 1024×768 96ppi
monitor, and scale that into `vh`s.