Body text rhythm
We have a line height of 1.5. To completely preserve line rhythm, we
could:

* separate paragraphs with a margin of 1.5em; *or*
* indent the first line of each paragraph and do away with the margin.

Neither is satisfactory for presentations — 1.5em is too wide and
indenting plays poorly with lists.

So we choose to use a margin of 0.75em. This way, the rhythm breaks on
odd paragraphs, and is restored on even ones.