Special classes for image slides
The markdown syntax for an image has a field for alt text. Since we turn
the image into a background, there’s nowhere we can put the alt text.
But we can (ab)use it to assign a special class to the slide, which then
can be used by the presentation author to style text of the slide.